Add compression support, make all data keys expire, update README, minor tweaks.
If allkeys-lru was used and maxmemory was reached it is possible for tag data to
get evicted since it is only "touched" on save, not load. If volatile-lru was used
it was possible for non-volatile keys to push out *all* volatile keys. So, now
all data keys are made volatile and all tag sets are non-volatile so that the
volatile-lru policy can be used with no side effects as long as maxmemory is sufficient.