[FIX] Make redis cache backend work like builtin cache
Setting lifetime to 0 evaluates to falsey, causing the infinite time
flag to be set to true. It also causes the expire to not be set.
According to docblock infinity should only be set iff lifetime is null.
This makes that work as expected and like magento cache works by
default.